name changes/action fix for visibility
